Post-Alamo Updates
Changes made before and during the Alamo regional. This update reflects
major chages in the robot design, namely
Post-Alamo Updates
Changes made before and during the Alamo regional. This update reflects
major changes made to the robot's design, namely the decision to move the
can lift from the Cube to the Pallet Jack. Some changes:
~ Cube's state machinge is simplified - now it only stacks. It does not
worry about tote count. The driver has a button which causes the clicker
to hover at the bottom.
~ Can lift and Claw added (part of same class, should be split).
~ Documentation - some small but nice things added, including a page of
joystick layouts and another with CAN id's. The main page has some links.
Comments for the main page are in RhsRobot.h